Added many more sensors to calltrack, changed return signature for sensors to return longs rather than ints

+++ include/fdb/fddb.h
@@ -160,7 +160,7 @@ FD_EXPORT u8_mutex fd_ipeval_lock;
 
 /* Cache calls */
 
-FD_EXPORT int fd_callcache_load(void);
+FD_EXPORT long fd_callcache_load(void);
 FD_EXPORT fdtype fd_cachecall(fdtype fcn,int n,fdtype *args);
 FD_EXPORT fdtype fd_xcachecall
   (struct FD_HASHTABLE *cache,fdtype fcn,int n,fdtype *args);
